tests[object]: Fix test case

+++ tests/unit-tests/object-manager.t.cpp
@@ -1,6 +1,6 @@
 /* -*- Mode:C++; c-file-style:"gnu"; indent-tabs-mode:nil; -*- */
 /**
- * Copyright (c) 2013-2016, Regents of the University of California.
+ * Copyright (c) 2013-2017, Regents of the University of California.
  *
  * This file is part of ChronoShare, a decentralized file sharing application over NDN.
  *
@@ -18,57 +18,53 @@
  * See AUTHORS.md for complete list of ChronoShare authors and contributors.
  */
 
-#include "logging.hpp"
 #include "object-manager.hpp"
 
-#include <boost/filesystem.hpp>
-#include <boost/filesystem/fstream.hpp>
+#include "test-common.hpp"
 
-#include <boost/make_shared.hpp>
-#include <boost/test/unit_test.hpp>
-#include <iostream>
-#include <iterator>
-#include <unistd.h>
+namespace ndn {
+namespace chronoshare {
+namespace tests {
 
-INIT_LOGGER("Test.ObjectManager");
-
-using namespace Ndnx;
-using namespace std;
-using namespace boost;
 namespace fs = boost::filesystem;
 
-BOOST_AUTO_TEST_SUITE(TestObjectManager)
+INIT_LOGGER("Test.ObjectManager")
 
-BOOST_AUTO_TEST_CASE(ObjectManagerTest)
+BOOST_FIXTURE_TEST_SUITE(TestObjectManager, IdentityManagementTimeFixture)
+
+BOOST_AUTO_TEST_CASE(FromFileToFile)
 {
-  INIT_LOGGERS();
+  fs::path tmpdir = fs::unique_path(UNIT_TEST_CONFIG_PATH) / "TestObjectManager";
+  if (exists(tmpdir)) {
+    remove_all(tmpdir);
+  }
 
-  fs::path tmpdir = fs::unique_path(fs::temp_directory_path() / "%%%%-%%%%-%%%%-%%%%");
-  _LOG_DEBUG("tmpdir: " << tmpdir);
   Name deviceName("/device");
 
-  CcnxWrapperPtr ccnx = make_shared<CcnxWrapper>();
-  ObjectManager manager(ccnx, tmpdir, "test-chronoshare");
+  shared_ptr<Face> face = make_shared<Face>(m_io);
 
-  tuple<HashPtr, int> hash_semgents =
-    manager.localFileToObjects(fs::path("test") / "test-object-manager.cc", deviceName);
+  ObjectManager manager(*face, tmpdir, "test-chronoshare");
 
-  BOOST_CHECK_EQUAL(hash_semgents.get<1>(), 3);
+  auto hash_segments =
+    manager.localFileToObjects(fs::path("tests") / "unit-tests" / "object-manager.t.cpp", deviceName);
 
-  bool ok = manager.objectsToLocalFile(deviceName, *hash_semgents.get<0>(), tmpdir / "test.cc");
+  BOOST_CHECK_EQUAL(std::get<1>(hash_segments), 3);
+
+  bool ok = manager.objectsToLocalFile(deviceName, *std::get<0>(hash_segments), tmpdir / "test.cpp");
   BOOST_CHECK_EQUAL(ok, true);
 
   {
-    fs::ifstream origFile(fs::path("test") / "test-object-manager.cc");
-    fs::ifstream newFile(tmpdir / "test.cc");
+    fs::ifstream origFile(fs::path("tests") / "unit-tests" / "object-manager.t.cpp");
+    fs::ifstream newFile(tmpdir / "test.cpp");
 
-    istream_iterator<char> eof, origFileI(origFile), newFileI(newFile);
+    std::istream_iterator<char> eof, origFileI(origFile), newFileI(newFile);
 
     BOOST_CHECK_EQUAL_COLLECTIONS(origFileI, eof, newFileI, eof);
   }
-
-  remove_all(tmpdir);
 }
 
-
 BOOST_AUTO_TEST_SUITE_END()
+
+} // namespace tests
+} // namespace chronoshare
+} // namespace ndn
